backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
SQLAlchemy==2.0.36
greenlet>=3.0.0
aiosqlite==0.20.0
pydantic==2.9.0
python-dotenv==1.0.1
python-dateutil==2.9.0.post0
uvicorn-worker==0.2.0
